requirements.txt
python-telegram-bot==20.3
openai==0.27.8
gspread==5.9.0
oauth2client==4.1.3
python-dotenv==1.0.0
